Use non-default Vite `port` and force to match `origin`.
Before this change, if you ran `run-assets` while another Vite server was
running somewhere on your local development system on also on the default port,
the new Vite server would automatically pick another free port, which would
then not match the origin configured and assets wouldn't load.

This would happen if trying to run `job-server` alongside `opencodelists`, for example.

Using port `8001` as this one more than `8000` used by the `just run`
webserver, so it's in the same "namespace".

Some templates and settings have the port hardcoded in dev mode. Change those and
therefore use `strictPort` as it's required for those to work.