Update base-docker to 24.04
This upgrades to newer Ubuntu and gets us a newer version of the
PostgreSQL client which should be backwards compatible with our existing
database, while allowing us to upgrade to PostgreSQL 16 in future.

This avoids the additional step of having to use the PostgreSQL
repository with Ubuntu 22.04 to get a suitable PostgreSQL client.