Handle rate limit for unauthenticated request test
We were seeing failures for job-server's verification tests more
frequently recently.

It's easy when reading the logs to think that this is a rate limit
issue on unauthenticated requests.

It *is*. But it also isn't.

This test specifically doesn't use the PAT that we use for testing. That means
we're:

* subject to GitHub's low rate limits for unauthenticated users
* and running in GitHub Actions runners that probably often end up using
  GitHub's API frequently

Those two issues combined mean we often exceed the rate limit.

Our test in fact is supposed to handle the case where we're rate limited
by GitHub, and pass successfully instead.

The problem is that our code doesn't quite handle the wrapping of the exception
properly. It treats our custom `HTTPError` exception as the `requests`
exception, tries to access the status code and fails.