Fix verification test failures introduced in #5184
The GitHub Issues API endpoint doesn't raise a 404 error if an issue
cannot be found, but instead the response dict's `total_count` == 0.

This fixes a verification test that expects a `IssueNotFound` to be
raised when no issues matching the given title are found.
Prior to this, the non-existence of the issue was not handled correctly
and subsequent API calls would try to append a comment to a nonexistent
issue, resulting in an unhandled HTTP 404 error.