Change the honeycomb queries to query all datasets
With the controller/agent split, we renamed the honeycomb datasets, but
the honeycomb links were specifically using the old jobrunner dataset.

Instead, run the query across all datasets, so we don't need to worry
about the name, and a resilient to future dataset/service name changes.

I have tested these new links manually, as there's not easy way to have
a functional test of them in CI.