Add defaults for new RAP settings
Add a default for new RAP settings. This is set to an empty string
as we expect it always to be set from environment variables, both in
dev and prod, and we don't need/want it to be a valid default. The
default ensures that the Airlock dev integration continues to work
and doesn't error due to missing env variables that it doesn't use.