Allow cancellation of a job request with cancelled jobs
This allows users to cancel an entire job request as long as there
are still some cancellable actions (actions that are pending/running
and haven't already been requested for cancellation).

Previously, if a single job in a job request was cancelled, the cancel
button for the entire job request was disabled.

Cancelling a single job just adds it to the job request's cancelled
actions. Cancelling an entire job request adds all active actions to
the job request's cancelled actions, so this doesn't cause any issues.

Also drive-by update to the slightly misleading button wording on the
job request page (Cancel this job -> Cancel this job request).