Add `rap_api_base_url` and `rap_api_token` fixtures.
This ensures that tests do not accidentally use the real API URL/token from an
environment variable via normal `settings.py`. Individual tests can use the
settings fixture to change the settings to some other value or blank.

This also reduces some boileplate from the tests.

Note that the value of the token doesn't matter in
`test_check_rap_api_status_bad_token`, so it's fine to use the same value.
That's because the response is mocked.