Add `jobserver.rap_api` module.
This encapsulates interactions with the RAP Controller via its API. This
reduces some repetition in the clients of the API and associated tests. Clients
of this module should not need to know about the HTTP layer or the
implementation details of the module.

Reference for how the server defines the API endpoints:
https://controller.opensafely.org/controller/v1/docs/
https://github.com/opensafely-core/job-runner/tree/main/controller/webapp/api_spec

Also updates the `check_rap_api_status` management command to use `jobserver.rap_api`.