Handle nothing to do in `request_cancellation`.
According to current view design (the clients of this function), it should
never happen that `request_cancellation` is called such that there is nothing
to do, as we check first. But nothing in this function precludes this and there
may be race conditions or future changes, so let's handle it for
belt-and-braces.