Refactor db_maintenance_mode() context processor
The context processor previously always called
Backend.objects.get_db_maintenance_mode_statuses(),
even if the current view was not one we'd
specified. This was wasteful and introduced side
effects.

This refactor restructures the context processor
so that get_db_maintenance_mode_statuses() is only
called if the current request’s URL name is one of
the ones we've specified.