Update hardcoded query numbers in tests
Certain tests which checked the number of db
queries being made previously failed, following
the addition of the `db_maintenance_mode()` context
processor which meant an additional qeury ran for
every test that requested a view.

Future work to refactor the context processor and future
add a fixture to override the `TEMPLATES`
setting, to stop the context processor running in
tests that don't need to care about it, mean that
the hardcoded numbers can be restored to what they
were originally.