Make `JobRequestCancel` extensible.
There is an almost-identical class-based view, `JobRequestCancel`, in the
`staff` views, and a very similar one, `JobCancel` in `jobserver`. Making these
use inheritance and overriding functions to represent only the differences will
make them easier to reason about and maintain and extend.

We could adopt a layered testing approach with the base case `post tested in
depth, the unique aspects of the derviced classes tested with unit tests, and
an integration test or two per view.

We want to extend each of these views to handle specific exceptions that may
arise with the extension of `JobRequest.request_cancellation` to call the RAP
API. It will be easier to do this if they have a common base.

The new class methods are conceptually private but I think it's more readable
if they aren't marked with a leading _, and not particularly confusing as
nothing in our application code calls any class methods directly.

Use `get_object_or_404` in `get_objects` rather than recreating that pattern
manually as was done before.