Add the ability to unpublish a snapshot
We add a management command for unpublishing a snapshot, which is
shorthand for retrospectively rejecting the most recent approved request
to publish a snapshot.

We use an existing method (`PublishRequest.reject()`) to do so, so most
of the management command is error handling. In common with most other
management commands, we exclude this management command from coverage.

Co-authored-by: Lucy Bridges <lucy.bridges@thedatalab.org>