Base `staff` `JobRequestCancel` on `jobserver` one.
These are almost identical.  Making these use inheritance and overriding
functions to represent only the differences will make them easier to reason
about and maintain and extend.

We could adopt a layered testing approach with the base case `post tested in
depth, the unique aspects of the derviced classes tested with unit tests, and
an integration test or two per view.

We want to extend each of these views to handle specific exceptions that may
arise with the extension of `JobRequest.request_cancellation` to call the RAP
API. It will be easier to do this if they have a common base.