Call RAP API when cancelling against `Test` backend.
We want requesting cancellation to be done via RAP API, rather than by just
setting `cancelled_actions` and waiting for the RAP API Controller sync loop.

Let's enable in the "Test" backend only initially so we can observe how that
behaves in production before rolling this out more fully.

I assert that because this behaviour is generic to the `JobRequestCancel` view
I need not edit those views derived from it. Given that, and as the new tests
of that view patch `JobRequest.request_cancellation`, I further assert that the
derived views don't need any new unit tests. We should redesign these tests to
use a layered approach in future to reduce the existing duplication. But that's
beyond the scope of this initiative.

TODO: logging and emit Sentry event. Intended for
https://github.com/opensafely-core/job-server/issues/5237
https://github.com/opensafely-core/job-server/issues/5238