Handle `request_cancellation` when no `Job` yet.
This can happen if it is called before we got any `Job` information from the
RAP side. Let's distinguish that from other cases.

This can be triggered in the staff pages by trying to cancel a job request
right after creating it. Separately, we might want to disable that button in
that case.