Update JobRequest status field when calculating job_status
The new status model field represents the overall status of a job
request. This may be a final state (all jobs succeeded) or it may
be an immediately set state for a job request with no jobs (nothing
to do, or failed on creating jobs).

We still need to calculate status from jobs, if the request has jobs
that are pending or running, but once the job request gets to a
completed state, we no longer need to calculate the status from
jobs again.