Handle RapAPIResponse/RequestErrors in create manage command
A RapAPIResponseError is expected to be returned when we receive a
response to a RAP API request, but the response indicates an error.
We should only raise this if we get a valid JSON response from the
endpoint.  For invalid json responses, we return a RapAPIRequestError,
indicating the request to the RAP api failed.

This means the code that calls ap_api.create() can safely mark the
job_request as failed if it encounters a RapAPIResponseError, and
can defer judgement to later update checks for a RapAPIRequestError,
which could just be a transient connection issue after jobs have
been created by the controller.

Any other exceptions raise by rap_api.create() are due to settings
errors, failure to contact the RAP api at all, or some uncaught
error in the rap_api module code itself, all of which are expected
to be outright failures.