Remove usage of Backend.parent_directory
Backend.parent_directory used to give the user a hint where to look in the filesystem on the backend when we used filesync.
It contains a string like D:\LEVEL4\workspaces. That message was removed in 89b4d75 when we pointed to airlock instead.
Thus the column and associated form handling are no longer needed.
Removing associated code only in this commit.