Remove Backend.parent_directory and migrate
Backend.parent_directory used to give the user a hint where to look in the filesystem on the backend when we usedfilesync. That message was removed in 89b4d75 when we pointed to airlock instead. Thus the column and associated form handling are no longer needed.

Safe to delete this column as all references to this field was removed in ac364f4.