Add a job status to test jobs
Jobs are created using job data sent from job-runner. This should
always have a job status. Now that we've reversed the filtering logic
for get_active_actions to filter for "pending" and "running" jobs rather
than excluding "succeeded" and "failed", we need to ensure that tests
for cancelling are set up with jobs with expected statuses.