Filter job requests that are considered active on the test backend
In the JobRequestAPIList, update the criteria by which we consider
a job request on the test backend to be active. This is the view that
is called by the (soon to be obsolete) controller sync loop.

Job requests on the test backend now call the RAP API when they are
created, and are acknowledged immediately. The RAP API does not create
dummy "failed" or "nothing to do" jobs, instead it sets the
job_request's status immediately. So a job_request for the test backend
can be already failed/nothing to do with 0 jobs.

For the test backend only, we now filter "active" job requests by the
job_request status AND job status. We'd expect these to be consistent,
so the filter by job status is just a safety check. We also include
job requests whose status is "unknown"; this should only ever happen if
we encountered some unhandled error whilst contacting the RAP API to create
jobs. In this case, we want to try again to retrieve the status of the
jobs from the controller.