Rename JobRequest.status to _status
We should always use jobs_status to get the overall status of a
job request. This will return the value of _status if the job_request
is complete already, or will calculate it from the status of jobs if
necessary. Rename it with a leading _ to indicate that it should be
considered private.