Update job-requests api's definition of active job request
Remove the conditional around the filtering of job requests in the
JobRequest API (called by the controller sync loop to retrieve job
requests that job server thinks are active). Having no jobs is now
a possible state for a completed job request, so these are not
included in the "active" requests.