Add failing tests for missing `user`.
The request object can be missing the user, we should test what happens in that case.

https://ebm-datalab.sentry.io/issues/6838552606/?project=5443358

Searching in Sentry shows a number of similar-looking issues. Seems the auth
middlewares are not always attaching user - maybe because of a database or
remote failure. The context processor should be made tolerant of that because
it's applied to every request - maybe including redirects or error pages.