Log an error for jobs that exist locally but not in controller
Previously we deleted the local jobs, now we just log an error.

If this happens, something's gone wrong somewhere. Jobs are only
created on job-server if the controller says they exist. We've seen
this happen before when a second controller was accidentally deployed
and calling job-server - because the second controller had no jobs,
job-server flipped between creating and deleting the jobs, depending
on which controller had tried to update it.  That particular scenario
shouldn't happen again, because of the change in direction of
communication.

Since this scenario is only expected in the event of an unexpected
error, let's not do anything as nuclear as deleting jobs.