Use update_or_create to update jobs
update_or_create uses select_for_update behind the scenes; it's
useful to use this at least at this stage in case something else
(i.e. the old, still-running controller sync loop) also attempts
to update this job at the same time.

This does involve a couple of additional db queries per job. We
expect this to go away when the RAP API V2 work is done and we
retire the controller sync loop and the update in the api/jobs
endpoint, and it could be removed. Future work may also want to
look into using bulk_create (with update_conflicts=True to also
handle updates)