Filter active jobs by test backend and fix some tests
We need to use the backend slug to filter active job requests that
look up by job status, otherwise we will actually be running these
updates against all backends, not just the test one.

Add a temporary helper function for creating a test backend with
the expected matching slug ("test") - as the BackendFactory uses an
incrementer by default, so the default slug is "test-0" which is
excluded.