Use a run_fn for the status update loop and move error handling
Using a run_fn (which by default just returns True) means that we can
test a finite number of loops.

Move the try/except inside the while loop so that a caught exception
doesn't cause the whole loop to stop. We want to instead catch, log and
keep the loop going.