Add a specific job request status for rap api request errors
This status means we can identify job requests where we
tried to create jobs using the RAP API and got an unhandled error,
which means we don't know if jobs were created or not. When the
status update code retrieves jobs from the controller, it will
either appropriately update this job request, OR if it finds no
jobs, it means it failed to create them and we can mark it as failed.

Any job requests that remain as UNKNOWN can only be so if their jobs
have weird and unrecognised statuses.