Make `Backend.level_4_url` nullable.
We plan to remove this field. We intend to remove the field from the model in
one PR then commit the migration in a separate PR to avoid issues on deployment
(per DEVELOPERS.md). We need to make the field nullable first so that the
objects created by the test framework are valid according to the database
schema from the migrations.

This is a safe migration to commit alongside the field change as making a field
nullable cannot cause a database validation error.