Update `_status` flag indicating `NotStartedYet`.
The semantics were updated in
https://github.com/opensafely-core/job-server/commit/f092e76f10200d9b18cde60752ec1cb24bdc2d3e and https://github.com/opensafely-core/job-server/commit/0061adf53e27a679db03dace43f7ad9001457098.

`JobRequest` with no jobs should be in either one of these statuses or
`NOTHING_TO_DO` after `request_rap_creation`.  This is the corrsponding change
to the cancel function.

Ignoring any possible `JobRequest` created any other way in this PR. Possibly
that is for https://github.com/opensafely-core/job-server/issues/5275.