Re-name `check_rap_api_status` to `rap_update_backend_status`.
This name was originally used when this was just a stub contacting a
healthcheck endpoint and is no longer appropriate. It updates the status of
each `Backend` model using the RAP API [backene status
endpoint](https://controller.opensafely.org/controller/v1/docs/#tag/backends).
This name uses the same prefix as the `rap_status` command for consistency.