Add example NDOO permissions file
Adds a draft permissions file for NDOO. This is in the current
permissions directory, in a sub directory named "population_permissions",
following the proposed categorisation here:
https://github.com/opensafely-core/job-runner/issues/1257#issuecomment-3559286804

The permissions directory may want to be updated to "analysis_scope",
"execution_config" or some other name that we decide on to encapsulate
permissions as well as access to features and other config that isn't
strictly permissions. However, I'll leave that for a later PR. If we
want to structure the permissions directory in line with the RAP API
categorisation, we'll probably also want to move the other files in
this directory into their respective category sub-directories.