Exclude population_permissions/ from IG ownership in CODEOWNERS
All jobserver/permissions/ files are owned by IG EXCEPT those in
jobserver/permissions/population_permissions. This is currently
just the ndoo permissions, which we do not expect to be owned by
IG, but will likely be changeable by devs with permission from the
OS service team. The exact process, and whether there's a team to
assign ownership to, is TBD. For now, any user with write access
can change files in jobserver/permissions/population_permissions/.