Remove create_label github api.
This was added for airlock issues, but the permissions job-server has
does not allow for the creation of labels in the target repo.

Instead, to avoid failure to create the issue, we filter out invalid
labels that don't exist. Hwoever, if any are missing, we warn via sentry
that they need creating.

I was somewhat suprised to find that there is no longer machinery for
asserting the content sentry messages emited during tests, so that part
is not tested.