Update virtualenv creation to use `uv venv`
This work is part of our migration to `uv`, and is
a stopgap towards the fuller uv workflow used in
our other services (i.e., `uv sync` + `uv.lock`).

We no longer allow users to specify a python
version and we don't need to specify a version in
the `uv` command to create a virtualenv; `uv`
checks the `.python-version` file
and uses the version stated there
(https://docs.astral.sh/uv/concepts/python-versions/#python-version-files).

We expect all devs to have `uv` installed locally
on path following announcement in the #tech Slack
channel and previous updates to DEVELOPERS.md.

Testing:
- Ran `just clean` then `just devenv`, which
  ultimately calls the updated `virtualenv`
  recipe (tried with `.python-version` set to
  3.12, then 3.13, to check behaviour of uv).
- Activated the newly-created virtualenv with
  dependencies installed
- Ran `just test-ci` and `just run` to confirm
  tests and local dev server ran successfully.