Add the full list of NDOO-exempt projects, as of 2025-12-05
Projects operating under the OpenSAFELY COVID service are out of scope
for National Data Opt-Outs.

Projects before 156 cannot be re-run, so are not included in the list.

This list consists of approved projects after #157, as of 2025-12-05 (#157-200);
all of these are COVID projects and are out of scope for NDOO.

See discussion in https://bennettoxford.slack.com/archives/C069YDR4NCA/p1763629225884149