Add `ServiceAdministrator` role and fixture.
This role will eventually allow someone to create projects without an
associated application, add researchers to the project, and assign project-level
roles to those researchers.

This is part of `initiative:manual-projects`, and a step towards a general
effort to disaggregate `StaffAreaAdministrator`.

This role should not be assigned in production until we are handing over this initiative.

Add a fixture as tests may want to use this role once it has more than
one permission. Often, a `UserFactory` for the specific relevant permissions
will make the test more precise and less brittle, though.