Test dokku deployments locally
This sets up dokku using docker, to run on a dev machine, so that changes to
the deployment can more easily be tested.

This currently doesn't work. It doesn't seem to be passing the env vars into
the running container correctly and it's quite likely the database hasn't been
setup correctly.

Test using `just local-deploy`.

Based on https://github.com/opensafely-core/opencodelists/pull/2774

Refs #4994